third_party/WebKit/LayoutTests/external/wpt/dom/lists/DOMTokenList-iteration-expected.txt

This is a testharness.js-based test.
FAIL DOMTokenList iteration: keys, values, etc. assert_equals: expected function "function keys() { [native code] }" but got function "function keys() { [native code] }"
Harness: the test ran to completion.
